send email in job, remove user's first name
