+++ app/views/user_mailer/in_person_verified.html.erb
@@ -8,7 +8,7 @@
 <h4><%= message.subject %></h4>
 <p>
 <div class="s16">
-  <%= t('user_mailer.in_person_verified.greeting', name: @first_name) %><br>
+  <%= t('user_mailer.in_person_verified.greeting') %><br>
   <%= t('user_mailer.in_person_verified.intro', location: @presenter.location, date: @presenter.date) %><br /><br />
   <%= t('user_mailer.in_person_verified.next_sign_in') %>
 </div>
